requirements.txt
flask
gunicorn
requests
python-dotenv
huggingface_hub